+++ Sage2020ViewDocListener.h
@@ -12,14 +12,14 @@ class Sage2020ViewDocListener {
   }
 
   virtual void DocEditNotification(int iLine, int cLine) = 0;
-  virtual void DocVersionChangedNotification(int nVer) = 0;
+  virtual void DocVersionChangedNotification(size_t nVer) = 0;
 
   Sage2020ViewDocListener* LinkListener(Sage2020ViewDocListener* listener_head);
   Sage2020ViewDocListener* UnlinkListener(
       Sage2020ViewDocListener* listener_head);
 
   void NotifyAllListenersOfEdit(int iLine, int cLine);
-  void NotifyAllListenersOfVersionChange(int nVer);
+  void NotifyAllListenersOfVersionChange(size_t nVer);
 
  protected:
   Sage2020ViewDocListener* next_listener_;
